producer/requirements.txt
kafka-python
yfinance

